# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Genrel Api",
        "description": "Apis for Goodreads Website!",
        "version": "1.0.0"
    },
    "servers": [{
        "url": "http://localhost:3000/"
    }],
    "security": [{
        "bearerAuth": []
    }],
    "paths": {
        "/user/signup": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create New User",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "genreList": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User Created"
                    },
                    "409": {
                        "description": "User All Ready Exist"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create New User",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Auth Sucessfull"
                    },
                    "401": {
                        "description": "Auth Fail"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/admin-user/{email}": {
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Give admin access to user",
                "parameters": [{
                    "name": "email",
                    "in": "path",
                    "description": "user email",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Admin added"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
        "/user/{email}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete a user.",
                "parameters": [{
                    "name": "email",
                    "in": "path",
                    "description": "user email",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "user Deleted"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
        "/user/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get All users",
                "description": "get all users",
                "responses": {
                    "200": {
                        "description": "all users"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
...
        "/review/": {
            "get": {
                "tags": ["Review"],
                "summary": "Get All reviews",
                "responses": {
                    "200": {
                        "description": "Get Reviews"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
        "/review/new": {
            "get": {
                "tags": ["Review"],
                "summary": "Get New reviews",
                "responses": {
                    "200": {
                        "description": "Get new Reviews"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
        "/review/user-book": {
            "get": {
                "tags": ["Review"],
                "summary": "Get All reviews of Single User",
                "responses": {
                    "200": {
                        "description": "Get Reviews"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
        "/review/book-review": {
            "post": {
                "tags": ["Review"],
                "summary": "Get All reviews of Single Book",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "bId": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get Reviews"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
        "/review/book-review-count": {
            "post": {
                "tags": ["Review"],
                "summary": "Get review Count Single Book",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "bId": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get Reviews"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
        "/review/r": {
            "post": {
                "tags": ["Review"],
                "summary": "Get reviews By Id",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "rId": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get Reviews"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
        "/review/add-review": {
            "post": {
                "tags": ["Review"],
                "summary": "create Review",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "read_id": {
                                        "type": "number"
                                    },
                                    "rating": {
                                        "type": "number"
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review Added"
                    },
                    "401": {
                        "description": "Auth Fail"
                    },
                    "409": {
                        "description": " First complete Readeing"
                    },
                    "400": {
                        "description": "Read not Foind"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/review/edit-review/{id}": {
            "patch": {
                "tags": ["Review"],
                "summary": "Edit Review",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "review id",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "number"
                    }
                }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "rating": {
                                        "type": "number"
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review Edited"
                    },
                    "401": {
                        "description": "Auth Fail"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/review/delete/{id}": {
            "delete": {
                "tags": ["Review"],
                "summary": "delete Review",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "review id",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "number"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Review Deleted"
                    },
                    "401": {
                        "description": "Auth Fail"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}